corrections after bdbaddog's code review
