+++ src/engine/SCons/Script/Main.py
@@ -59,7 +59,7 @@ import SCons.Job
 import SCons.Node
 import SCons.Node.FS
 import SCons.Platform
-import SCons.Platform.VE
+import SCons.Platform.virtualenv
 import SCons.SConf
 import SCons.Script
 import SCons.Taskmaster
@@ -864,12 +864,12 @@ def _main(parser):
     for warning_type, message in delayed_warnings:
         SCons.Warnings.warn(warning_type, message)
 
-    if not SCons.Platform.VE.virtualenv_enabled_by_default:
+    if not SCons.Platform.virtualenv.virtualenv_enabled_by_default:
         if options.enable_virtualenv:
-            SCons.Platform.VE.enable_virtualenv = True
+            SCons.Platform.virtualenv.enable_virtualenv = True
 
     if options.ignore_virtualenv:
-        SCons.Platform.VE.ignore_virtualenv = True
+        SCons.Platform.virtualenv.ignore_virtualenv = True
 
     if options.diskcheck:
         SCons.Node.FS.set_diskcheck(options.diskcheck)
