+++ test/virtualenv/activated/virtualenv_activated_python.py
@@ -32,17 +32,17 @@ environment or in unactivated virtualenv.
 """
 
 import TestSCons
-import SCons.Platform.VE
+import SCons.Platform.virtualenv
 import sys
 import os
 import re
 
 test = TestSCons.TestSCons()
 
-if not SCons.Platform.VE.Virtualenv():
+if not SCons.Platform.virtualenv.Virtualenv():
     test.skip_test("No virtualenv detected, skipping\n")
 
-if not SCons.Platform.VE.select_paths_in_venv(os.getenv('PATH')):
+if not SCons.Platform.virtualenv.select_paths_in_venv(os.getenv('PATH')):
     test.skip_test("Virtualenv detected but looks like unactivated, skipping\n")
 
 
@@ -53,7 +53,7 @@ print("sys.executable: %s" % repr(sys.executable))
 print("env.WhereIs('python'): %s" % repr(env.WhereIs('python')))
 """)
 
-if SCons.Platform.VE.virtualenv_enabled_by_default:
+if SCons.Platform.virtualenv.virtualenv_enabled_by_default:
     test.run(['-Q'])
 else:
     test.run(['-Q', '--enable-virtualenv'])
@@ -82,9 +82,9 @@ can't determine env.WhereIs('python') from stdout:
 python = eval(m.group('py'))
 
 # runing in activated virtualenv (after "activate") - PATH includes virtualenv's bin directory
-test.fail_test(not SCons.Platform.VE.IsInVirtualenv(interpreter),
+test.fail_test(not SCons.Platform.virtualenv.IsInVirtualenv(interpreter),
                message="sys.executable points outside of virtualenv")
-test.fail_test(not SCons.Platform.VE.IsInVirtualenv(python),
+test.fail_test(not SCons.Platform.virtualenv.IsInVirtualenv(python),
                message="env.WhereIs('python') points outside of virtualenv")
 
 test.pass_test()
